Undo refactoring of FailableStream.
