+++ src/test/java/org/apache/commons/lang3/stream/TestConstants.java
@@ -24,24 +24,24 @@ import java.util.List;
  * {@link org.apache.commons.lang3.stream}.
  */
 public class TestConstants {
-    static final String EXPECTED_NFE_MESSAGE_INT = "For input string: \"4 \"";
-    static final String EXPECTED_NFE_MESSAGE_DOUBLE = "For input string: \"4.5x\"";
-    static final String EXPECTED_NFE_MESSAGE_LONG = "For input string: \"4000000000 \"";
+    public static final String EXPECTED_NFE_MESSAGE_INT = "For input string: \"4 \"";
+    public static final String EXPECTED_NFE_MESSAGE_DOUBLE = "For input string: \"4.5x\"";
+    public static final String EXPECTED_NFE_MESSAGE_LONG = "For input string: \"4000000000 \"";
 
     private static final List<String> EXTRA_INTS = Arrays.asList("7", "8");
-    static final List<String> INPUT_INT = Arrays.asList("1", "2", "3", "4", "5", "6");
-    static final List<String> FAILING_INPUT_INT = Arrays.asList("1", "2", "3", "4 ", "5", "6");
-    static final List<List<String>> FLAT_MAP_INPUT_INT = Arrays.asList(INPUT_INT, EXTRA_INTS);
-    static final List<List<String>> FAILING_FLAT_MAP_INPUT_INT = Arrays.asList(FAILING_INPUT_INT,
+    public static final List<String> INPUT_INT = Arrays.asList("1", "2", "3", "4", "5", "6");
+    public static final List<String> FAILING_INPUT_INT = Arrays.asList("1", "2", "3", "4 ", "5", "6");
+    public static final List<List<String>> FLAT_MAP_INPUT_INT = Arrays.asList(INPUT_INT, EXTRA_INTS);
+    public static final List<List<String>> FAILING_FLAT_MAP_INPUT_INT = Arrays.asList(FAILING_INPUT_INT,
             EXTRA_INTS);
 
-    static final List<String> INPUT_LONG = Arrays.asList("1000000000", "2000000000", "3000000000",
+    public static final List<String> INPUT_LONG = Arrays.asList("1000000000", "2000000000", "3000000000",
             "4000000000", "5000000000", "6000000000");
-    static final List<List<String>> FLAT_MAP_INPUT_LONG = Arrays.asList(INPUT_LONG,
+    public static final List<List<String>> FLAT_MAP_INPUT_LONG = Arrays.asList(INPUT_LONG,
                     Arrays.asList("7000000000", "8000000000"));
 
-    static final List<String> INPUT_DOUBLE = Arrays.asList("1.5", "2.5", "3.5", "4.5", "5.5", "6.5");
-    static final List<List<String>> FLAT_MAP_INPUT_DOUBLE = Arrays.asList(INPUT_DOUBLE,
+    public static final List<String> INPUT_DOUBLE = Arrays.asList("1.5", "2.5", "3.5", "4.5", "5.5", "6.5");
+    public static final List<List<String>> FLAT_MAP_INPUT_DOUBLE = Arrays.asList(INPUT_DOUBLE,
             Arrays.asList("7.5", "8.5"));
 
     private TestConstants() {}
